requirements.txt
fastapi
uvicorn
qdrant-client
sentence-transformers
groq
python-dotenv